Rename Template classes
